requirements.txt
asgiref==3.7.2
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
icecream==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==10.2.0
platformdirs==4.1.0
Pygments==2.17.2
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
virtualenv==20.25.0
Werkzeug==3.0.1
gunicorn==20.1.0